plugin-EnergyCampus: refactoring
